Extract code to automatically derive data from maven-project
Currently if one loads the configuration with
aQute.bnd.maven.lib.configuration.BndConfiguration then only the
configuration itself is loaded, but actually some implicit configuration
later apply (e.g. from the maven project configuration)
that is currently only performed by the AbstractBndMavenPlugin.

This extracts this logic from the AbstractBndMavenPlugin into the
BndConfiguration class for the following benefits:

- making the AbstractBndMavenPlugin maintain less code in the execute
method
- allow other users of BndConfiguration to easily inherit common
defaults
- separation of concerns, easier to debug and understand
